Download functionality is basically working
It's still a bit problematic, as it will redownload the same images that
you already have had, but at least the base functionality works for now.

+++ src/config.rs
@@ -23,7 +23,7 @@ pub struct Tapet {
 
 #[derive(Debug, Deserialize)]
 pub struct Wallhaven {
-    pub download_number: u64,
+    pub download_number: u32,
     pub tags: String,
 }
